generify UniProtKBSearchFields, fix todos, and add more tests
