+++ uniprot-search/src/main/java/org/uniprot/store/search/domain2/SearchItem.java
@@ -34,8 +34,6 @@ public interface SearchItem {
 
     String getDescription();
 
-    String getSites();
-
     String getExample();
 
     String getItemType();
