+++ uniprot-search/src/test/resources/uniprot/search-fields.json
@@ -48,7 +48,7 @@
             "label": "NOTE",
             "field": "cc_cofactor_note",
             "evidenceField": "ccev_cofactor_note",
-            "rangeField" : "pretend_cofactor_range_field",
+            "rangeField" : "pretend_range_field",
             "dataType": "string",
             "description": "Search by cofactor note",
             "example": "subunit"
